pyproject.toml
[project]
name = "stockreport"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.15",
    "langchain-anthropic>=1.4.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.26",
    "langchain-google-genai>=4.2.1",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.1.12",
    "langchain-pinecone>=0.2.13",
    "langchain-text-splitters>=1.1.1",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
    "tabulate>=0.10.0",
    "yfinance>=1.2.0",
]
